ajusts in game service

+++ magic-repository/src/main/kotlin/com/magic/repository/CardRepository.kt
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CardRepository: JpaRepository<Card,Long> {
+interface CardRepository: JpaRepository<Card,Int> {
     fun findByPlayerType(playerType: PlayerType): List<Card>?
 }
