+++ magic-repository/src/main/kotlin/com/magic/repository/GameRepository.kt
@@ -3,5 +3,5 @@ package com.magic.repository
 import com.magic.model.Game
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface GameRepository: JpaRepository<Game, Long> {
+interface GameRepository: JpaRepository<Game, Int> {
 }
