+++ magic-repository/src/main/kotlin/com/magic/repository/PlayerRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PlayerRepository : JpaRepository<Player, Long> {
+interface PlayerRepository : JpaRepository<Player, Int> {
 }
